Move class definition into domain layer
