+++ OTAnalytics/adapter_visualization/color_provider.py
@@ -1,7 +1,7 @@
 import random
 
-from OTAnalytics.adapter_visualization.otc_classes import OtcClasses
-from OTAnalytics.adapter_visualization.tls8plus1_classes import Tls8plus1Classes
+from OTAnalytics.domain.otc_classes import OtcClasses
+from OTAnalytics.domain.tls8plus1_classes import Tls8plus1Classes
 
 DEFAULT_COLOR_PALETTE: dict[str, str] = {
     OtcClasses.CAR: "blue",
